# app/receiver/service/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /v1/business/create_app:
        post:
            tags:
                - Business
            operationId: Business_CreateAPP
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/server.v1.CreateAPPRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/server.v1.CreateAPPResponse'
    /v1/business/create_client:
        post:
            tags:
                - Business
            operationId: Business_CreateClient
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/server.v1.CreateClientRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/server.v1.CreateClientResponse'
    /v1/business/get_app:
        post:
            tags:
                - Business
            operationId: Business_GetAPP
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/server.v1.GetAPPRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/server.v1.GetAPPResponse'
    /v1/business/get_client:
        post:
            tags:
                - Business
            operationId: Business_GetClient
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/server.v1.GetClientRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/server.v1.GetClientResponse'
    /v1/business/get_client_list:
        post:
            tags:
                - Business
            operationId: Business_GetClientList
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/server.v1.GetClientListRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/server.v1.GetClientListResponse'
    /v1/message/set_message:
        post:
            tags:
                - Message
            operationId: Message_SendMessage
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/server.v1.SendMessageRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/server.v1.SendMessageResponse'
components:
    schemas:
        server.v1.ClientInfo:
            type: object
            properties:
                clientId:
                    type: integer
                    format: int32
                clientName:
                    type: string
        server.v1.CreateAPPRequest:
            type: object
            properties:
                appName:
                    type: string
                appKey:
                    type: string
        server.v1.CreateAPPResponse:
            type: object
            properties:
                appId:
                    type: integer
                    format: int32
        server.v1.CreateClientRequest:
            type: object
            properties:
                clientName:
                    type: string
                clientKey:
                    type: string
                appId:
                    type: integer
                    format: int32
        server.v1.CreateClientResponse:
            type: object
            properties:
                clientId:
                    type: integer
                    format: int32
        server.v1.GetAPPRequest:
            type: object
            properties:
                appId:
                    type: integer
                    format: int32
        server.v1.GetAPPResponse:
            type: object
            properties:
                appId:
                    type: integer
                    format: int32
                appName:
                    type: string
        server.v1.GetClientListRequest:
            type: object
            properties:
                appId:
                    type: integer
                    format: int32
        server.v1.GetClientListResponse:
            type: object
            properties:
                clients:
                    type: array
                    items:
                        $ref: '#/components/schemas/server.v1.ClientInfo'
        server.v1.GetClientRequest:
            type: object
            properties:
                clientId:
                    type: integer
                    format: int32
        server.v1.GetClientResponse:
            type: object
            properties:
                client:
                    $ref: '#/components/schemas/server.v1.ClientInfo'
        server.v1.SendMessageRequest:
            type: object
            properties:
                appId:
                    type: integer
                    format: int32
                clientIds:
                    type: array
                    items:
                        type: integer
                        format: int32
                content:
                    type: string
                delay:
                    type: string
                sendTime:
                    type: string
                sendCount:
                    type: integer
                    format: int32
        server.v1.SendMessageResponse:
            type: object
            properties:
                messageId:
                    type: integer
                    format: int32
tags:
    - name: Business
    - name: Message